requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
numpy
python-multipart
python-dotenv
pydantic-settings